update some confgiration file
