+++ .eslintrc
@@ -1,17 +1,59 @@
 {
-  "root": true,
-  "parser": "@typescript-eslint/parser",
-  "plugins": [
-	"@typescript-eslint"
-  ],
-  "extends": [
-	"eslint:recommended",
-	"plugin:@typescript-eslint/eslint-recommended",
-	"plugin:@typescript-eslint/recommended"
-  ],
-  "rules": {
-	"@typescript-eslint/no-useless-constructor": "off",
-	"@typescript-eslint/no-empty-function": "off",
-	"@typescript-eslint/no-explicit-any": "off"
-  }
+    "env": {
+        "jquery": true,
+        "node": true,
+        "browser": true,
+        "es2020": true,
+        "es6": true
+    },
+    "root": true,
+    "parser": "@typescript-eslint/parser",
+    "parserOptions": {
+        "ecmaVersion": 2020,
+        "sourceType": "module"
+    },
+    "extends": [
+        "eslint:recommended",
+        "@typhonjs-fvtt/eslint-config-foundry.js/0.8.0",
+        "plugin:@typescript-eslint/recommended",
+        "plugin:prettier/recommended"
+    ],
+    "plugins": [
+        "@typescript-eslint",
+        "eslint-plugin-tsdoc"
+    ],
+    "rules": {
+        "@typescript-eslint/no-useless-constructor": "off",
+        "@typescript-eslint/no-empty-function": "off",
+        "@typescript-eslint/no-explicit-any": "off",
+        "tsdoc/syntax": "off",
+        "@typescript-eslint/ban-types": "off",
+        "prettier/prettier": [
+            "off",
+            {},
+            {
+              "usePrettierrc": true,
+            }
+        ],
+        "no-underscore-dangle": "off",
+        "import/extensions": "off",
+        "class-methods-use-this": [
+            "off",
+            {
+                "exceptMethods": ["getData", "_updateObject"]
+            }
+        ],
+        "@typescript-eslint/consistent-type-assertions": "off",
+        "@typescript-eslint/ban-ts-comment": "off",
+        "no-shadow": ["off", { "builtinGlobals": true, "hoist": "all", "allow": ["event"] }],
+        "@typescript-eslint/explicit-function-return-type": "off",
+        "@typescript-eslint/ban-ts-ignore": "off",
+        "@typescript-eslint/no-this-alias": [
+            "error",
+            {
+              "allowDestructuring": false, // Disallow `const { props, state } = this`; true by default
+              "allowedNames": ["self"] // Allow `const self = this`; `[]` by default
+            }
+          ]
+    }
 }
