+++ .prettierignore
@@ -1,0 +1,6 @@
+/dist
+/package-lock.json
+/.pnp.js
+/.yarn/
+/.vscode/
+/.idea
