+++ package.json
@@ -1,19 +1,23 @@
 {
+  "private": true,
 	"name": "merchantsheetnpc",
+  "title": "",
+  "description": "",
 	"version": "2.6.0",
-	"description": "",
 	"main": "index.js",
-	"scripts": {
+  "scripts": {
 		"package": "gulp publish --update",
 		"build": "gulp build --dbg && gulp link && npm test",
 		"build:prod": "gulp build && npm test",
 		"build:watch": "gulp watch",
 		"clean": "gulp clean && gulp link --clean",
-		"test": "jest"
-	},
+		"test": "jest",
+    "update": "npm install --save-dev @league-of-foundry-developers/foundry-vtt-types@9",
+    "prettier-format": "prettier --config .prettierrc --write \"./Source/**/*.(ts|js|json|scss|css|html)\""
+  },
 	"author": "whelan",
 	"license": "GPLv3",
-	"devDependencies": {
+  "devDependencies": {
 		"@babel/core": "^7.15.0",
 		"@league-of-foundry-developers/foundry-vtt-types": "^9.242.0",
 		"@types/archiver": "^5.1.1",
@@ -35,8 +39,13 @@
 		"gulp-git": "^2.10.1",
 		"gulp-less": "^5.0.0",
 		"gulp-sourcemaps": "^3.0.0",
-		"gulp-typescript": "^5.0.0",
 		"gulp-uglify": "^3.0.2",
+    "gulp-changed": "^4.0.3",
+    "gulp-eslint": "^6.0.0",
+    "gulp-rename": "^2.0.0",
+    "gulp-replace": "^1.1.3",
+    "gulp-sass": "^4.1.1",
+    "gulp-typescript": "^6.0.0-alpha.1",
 		"jest": "^27.5.1",
 		"json-stringify-pretty-compact": "^3.0.0",
 		"ts-jest": "^27.1.3",
@@ -46,10 +55,14 @@
 		"vinyl-buffer": "^1.0.1",
 		"vinyl-source-stream": "^2.0.0",
 		"yargs": "^17.1.1"
-	},
-	"dependencies": {
-		"csv-parse": "^4.16.3",
+  },
+  "dependencies": {
+    "build": "^0.1.4",
+    "node-sass": "^4.14.1",
+    "prettier-format": "^3.1.0",
+    "typescript": "^4.6.3",
+    "csv-parse": "^4.16.3",
 		"gulp-jest": "^4.0.4",
 		"tsify": "^5.0.4"
-	}
+  }
 }
